Add files via upload
This code is basically for driver's drowsiness detection. To detect drowsiness we will be considering two factors one is frequency of blinks and other being yawn. These two factors are defined by respect to aspect ratios. When the given threshold conditions  are met, drowsiness is detected and the driver is alarmed.